[refactor][볼링-step3] 볼링 결과 상태 패키지 정리
